Task016-021 done.

+++ src/main/java/com/epam/university/java/core/task015/PointFactoryImpl.java
@@ -1,8 +1,10 @@
 package com.epam.university.java.core.task015;
 
 public class PointFactoryImpl implements PointFactory {
+
+
     @Override
-    public Point newInstance(int x, int y) {
+    public Point newInstance(double x, double y) {
         return new PointImpl(x, y);
     }
 }
